fix(headless): working analysis start and improved try/except

+++ hub_adapter/headless.py
@@ -4,6 +4,7 @@ import logging
 import uuid
 
 from fastapi import HTTPException
+from httpx import HTTPStatusError
 from starlette import status
 
 from hub_adapter.auth import get_internal_token
@@ -14,10 +15,16 @@ from hub_adapter.routers.hub import (
     compile_analysis_pod_data,
     format_query_params,
     get_node_id,
+    get_node_metadata_for_url,
     get_node_type,
+    get_registry_metadata_for_url,
     list_analysis_nodes,
 )
-from hub_adapter.routers.kong import create_and_connect_analysis_to_project, delete_analysis, list_projects
+from hub_adapter.routers.kong import (
+    create_and_connect_analysis_to_project,
+    delete_analysis,
+    list_projects,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -41,8 +48,12 @@ async def auto_start_analyses():
     node_id = await get_node_id()
     node_type = await get_node_type()
 
-    formatted_query_params = format_query_params({"sort": "-updated_at", "include": "analysis"})
-    analyses = await list_analysis_nodes(node_id=node_id, query_params=formatted_query_params)
+    formatted_query_params = format_query_params(
+        {"sort": "-updated_at", "include": "analysis"}
+    )
+    analyses = await list_analysis_nodes(
+        node_id=node_id, query_params=formatted_query_params
+    )
     valid_projects = await get_valid_projects()
     ready_to_start_analyses = parse_analyses(analyses, valid_projects)
 
@@ -64,42 +75,63 @@ async def auto_start_analyses():
             "node_id": node_id,
             "kong_token": kong_token,
         }
-        start_resp = await start_analysis_pod(analysis_props=props)
-        logger.info(f"Analysis start response for {analysis_id}: {start_resp['status']}")
+        await start_analysis_pod(analysis_props=props, kong_token=kong_token)
 
 
-async def register_analysis(analysis_id: str, project_id: str, attempt: int = 1, max_attempts: int = 5) -> dict | None:
+async def register_analysis(
+    analysis_id: str, project_id: str, attempt: int = 1, max_attempts: int = 5
+) -> dict | None:
     """Register an analysis with kong."""
     logger.info(f"Attempt {attempt} at starting analysis {analysis_id}")
     try:
-        kong_resp = await create_and_connect_analysis_to_project(project_id, analysis_id)
+        kong_resp = await create_and_connect_analysis_to_project(
+            project_id, analysis_id
+        )
         return kong_resp
 
     except HTTPException as e:
         if e.status_code == status.HTTP_404_NOT_FOUND:
-            logger.error(f"{e.detail['message']}, failed to start analysis {analysis_id}")
+            logger.error(
+                f"{e.detail['message']}, failed to start analysis {analysis_id}"
+            )
 
         elif e.status_code == status.HTTP_409_CONFLICT:
-            logger.warning(f"Analysis {analysis_id} already registered, checking if pod exists...")
+            logger.warning(
+                f"Analysis {analysis_id} already registered, checking if pod exists..."
+            )
             pod_exists = await pod_running(analysis_id)
-            if not pod_exists:  # If not running, then delete the kong consumer
-                logger.info(f"No pod found for {analysis_id}, will delete kong consumer and retry")
+            if pod_exists is None:  # Status could not be obtained, skip and try later
+                pass
+
+            elif (
+                not pod_exists
+            ):  # Status obtained and if not running, delete kong consumer
+                logger.info(
+                    f"No pod found for {analysis_id}, will delete kong consumer and retry"
+                )
                 await delete_analysis(analysis_id=analysis_id)
 
                 if attempt < max_attempts:
-                    return await register_analysis(analysis_id, project_id, attempt + 1, max_attempts)
+                    return await register_analysis(
+                        analysis_id, project_id, attempt + 1, max_attempts
+                    )
 
             else:
-                logger.info(f"Pod already exists for analysis {analysis_id}, skipping start sequence")
+                logger.info(
+                    f"Pod already exists for analysis {analysis_id}, skipping start sequence"
+                )
 
         else:
             logger.error(f"Failed to start analysis {analysis_id}, {e}")
 
 
-async def pod_running(analysis_id: str) -> bool:
+async def pod_running(analysis_id: str) -> bool | None:
     """Check whether a pod with the given analysis_id is already running."""
     pod_status = await fetch_analysis_status(analysis_id=analysis_id)
-    return bool(pod_status["status"])  # If anything exists
+    if pod_status:
+        return bool(pod_status["status"])  # If anything exists
+
+    return None  # Error occurred and no status retrieved
 
 
 async def fetch_token_header() -> dict:
@@ -109,35 +141,56 @@ async def fetch_token_header() -> dict:
     return token
 
 
-async def start_analysis_pod(analysis_props: dict) -> dict:
+async def start_analysis_pod(analysis_props: dict, kong_token: str):
     """Start a new analysis pod."""
     logger.info(f"Starting new analysis pod for {analysis_props['analysis_id']}")
 
-    props = compile_analysis_pod_data()
+    node_metadata = get_node_metadata_for_url(analysis_props["node_id"])
+    analysis_info = get_registry_metadata_for_url(node_metadata)
+    props = compile_analysis_pod_data(
+        analysis_id=analysis_props["analysis_id"],
+        project_id=analysis_props["project_id"],
+        compiled_info=analysis_info,
+        kong_token=kong_token,
+    )
 
     headers = await fetch_token_header()
     microsvc_path = f"{hub_adapter_settings.PODORC_SERVICE_URL}/po/"
 
-    resp_data, _ = await make_request(
-        url=microsvc_path,
-        method="post",
-        headers=headers,
-        data=analysis_props,
-    )
+    try:
+        resp_data, _ = await make_request(
+            url=microsvc_path,
+            method="post",
+            headers=headers,
+            data=props,
+        )
+        logger.info(
+            f"Analysis start response for {analysis_props["analysis_id"]}: {resp_data['status']}"
+        )
 
-    return resp_data
+    except HTTPStatusError as e:
+        logger.error(
+            f"Unable to start analysis {analysis_props['analysis_id']} due to the following error: {e}"
+        )
 
 
 async def fetch_analysis_status(analysis_id: uuid.UUID | str) -> dict:
     """Fetch the status for a specific analysis run. For headless operation"""
     headers = await fetch_token_header()
     microsvc_path = f"{hub_adapter_settings.PODORC_SERVICE_URL}/po/{analysis_id}/status"
+    resp_data = None
 
-    resp_data, _ = await make_request(
-        url=microsvc_path,
-        method="get",
-        headers=headers,
-    )
+    try:
+        resp_data, _ = await make_request(
+            url=microsvc_path,
+            method="get",
+            headers=headers,
+        )
+
+    except HTTPStatusError as e:
+        logger.error(
+            f"Unable to fetch the status of analysis {analysis_id} due to the following error: {e}"
+        )
 
     return resp_data
 
@@ -167,7 +220,12 @@ def parse_analyses(analyses: list, valid_projects: set) -> set:
             entry.run_status,
             entry.approval_status,
         )
-        if approved == "approved" and build_status == "finished" and not run_status and project_id in valid_projects:
+        if (
+            approved == "approved"
+            and build_status == "finished"
+            and not run_status
+            and project_id in valid_projects
+        ):
             valid_entry = (analysis_id, project_id, node_id, build_status, run_status)
             ready_analyses.add(valid_entry)
